+++ gdesk/panels/imgview/imgpaint.py
@@ -18,7 +18,7 @@ from ... import config, gui
 from qtpy import QtCore, QtGui, QtWidgets, API_NAME
 from qtpy.QtCore import Qt, Signal, QUrl
 from qtpy.QtGui import QFont, QTextCursor, QPainter, QPixmap, QCursor, QPalette, QColor, QKeySequence, QFontMetrics
-from qtpy.QtWidgets import QAction, QVBoxLayout, QWidget
+from qtpy.QtWidgets import QApplication, QWidget
 
 from ...panels import thisPanel
 
@@ -89,8 +89,12 @@ class ImageViewerWidget(QWidget):
         self.sel_pix_x = 0
         self.sel_pix_y = 0
 
-        self.setBackgroundColor(*config['image background'])
-        
+        color_scheme = QApplication.instance().color_scheme
+        if color_scheme == "Dark":
+            self.setBackgroundColor(*config["image background dark"])
+        else:
+            self.setBackgroundColor(*config["image background"])
+
         self.set_val_item_format('dec')        
 
         self.pickCursor = QCursor(QPixmap(str(respath / "icons" / "pixPick256.png")), 15, 15)